6 성공 long으로 수정

+++ b10844/src/Main.java
@@ -6,8 +6,8 @@ public class Main {
         BufferedWriter bw = new BufferedWriter(new OutputStreamWriter(System.out));
         int N = Integer.parseInt(bf.readLine());
 
-        int[] before = new int[10];
-        int[] after = new int[10];
+        long[] before = new long[10];
+        long[] after = new long[10];
         before[0] = 0;
         for(int i = 1; i <= 9; i++) {
             before[i] = 1;
@@ -16,14 +16,14 @@ public class Main {
         for(int i = 2; i <= N; i++) {
             fillCountArr(before, after);
             before = after;
-            after = new int[10];
+            after = new long[10];
         }
         
-        bw.write(Integer.toString(sumArrDiv(before) % 1000000000));
+        bw.write(Long.toString(sumArrDiv(before) % 1000000000));
         bw.close();
     }
 
-    public static void fillCountArr(int[] before, int[] after) {
+    public static void fillCountArr(long[] before, long[] after) {
         for(int i = 1; i <= 8; i++) {
             after[i] = (before[i - 1] + before[i + 1]) % 1000000000;
         }
@@ -31,8 +31,8 @@ public class Main {
         after[9] = before[8] % 1000000000;
     }
 
-    public static int sumArrDiv(int[] after) {
-        int sum = 0;
+    public static long sumArrDiv(long[] after) {
+        long sum = 0;
         for(int i = 0; i <= 9; i++) {
             sum += after[i];
         }
